job-info: Wait for guest eventlogs to be created
When watching eventlogs in the guest eventlog namespace, always
wait for their creation.  Unlike eventlogs in the primary namespace,
there can be races where the guest namespace has been created but
the eventlog has not yet been created.

In the event the eventlog is never created, behavior is still mimiced
identically, in that the watch will "hang" with no events ever
occurring.

Fixes #2346