docker: add a systest stage to the fluxorama Dockerfile
Problem: The fluxorama "system test" docker container is not ideal
for system testing since it includes and runs ttyd, unnecessarily
exposing an external port for web browser shell.

Add a 'systest' stage to the fluxorama Dockerfile which doesn't
have ttyd enabled. This stage will be used as part of automated
and interactive system testing outside of ttyd mechanism.

Since ttyd is not necessarily running, make the requirement for
FLUXUSER_PASSWORD dependent on the ttyd service in entrypoint.sh.