docker: add script to run in fluxorama docker container
Problem: The fluxorama docker image is not ideal for iterative
development.

Add a new script src/test/docker/docker-run-systest.sh, which,
similar to `docker-run-checks.sh -I`, starts a fluxorama docker
container with the current UID/GID registered, mounts the build
directory at /usr/src, and runs an interactive shell in the instance.

Later this script can be enhanced to run a testsuite under this
container for real-world multi-user testing.