python: use job-list "all" attr
Problem: The python bindings have a VALID_ATTRS list that is
maintained to grab all attrs from a job.  This is no longer
necessary with the job-list "all" attr.

Solution: Use the "all" attr and remove VALID_ATTRS.  As a side
effect, a copy of valid attrs must be added to tests when checking
for the list of legal attrs returned from job-list.list-attrs.