job-list: store copy of jobspec internally
Problem: In the near future we will need access to the job's
jobspec when a job goes inactive, but we do not store a copy
of it.

Solution: Store a reference to the jobspec in struct job.  As a
side effect, we no longer need to store several jobspec sub-objects
within struct job.