job-list: add job db
Problem: We would like to store inactive jobs on disk for retrieval
at a later time.

Solution: Add an sqlite job db to the job-list module.  If the
job-list module is configured with a database path or if the flux
broker is configured with a statedir, store all inactive job data
to the db.