job-manager: fix dependencies added in DEPEND state
Problem: Dependencies added during the job.state.depend callback
do not keep a job in DEPEND state, which is a change from previous
behavior. This occurs because the check for any pending dependencies
is done in event_job_action() while the "dependency-add" events are
still queued in job->event_queue. The job manager thus emits the
"depend" event, which transitions the job to SCHED state before the
queued dependency events can be applied.

Though an argument could be made that plugins can add dependencies in
either job.new or job.validate, since the ability to add dependencies
in job.state.depend is documented in flux-jobtap-plugins(7) and is in
use by at least two external projects, the previous behavior should
probably be preserved.

Add a check in event_job_action() for queued dependency-add events
before posting the "depend" event that transitions a job out of DEPEND
state. While this adds an extra check in the job state machine, the
common case will be that there are no queued events, and therefore
the check is likely to have little impact.

Fixes #4405