job-list: split out json_load() in jobspec_parse()
Problem: In the near future we will need to update the internally
cached jobspec and re-parse it.  The current jobspec_parse() function takes
the jobspec in raw string form from the KVS, making it unusable for
a re-parse.

Solution: Create a new function jobspec_load_and_parse(), that will
take the raw jobspec string and convert it into a cached json object.  Make
jobspec_parse() only to parse the stored jobspec object.