job-list: support constraint -> sql conversion
Problem: In the near future we would like to query jobs from the
job db.  In order to do so efficiently, we need to convert the
job list constraint into equivalent SQL that can be used in a
WHERE statement.

Support a new helper library "constraint_sql" that supports converting
a constraint into conditions that can be passed in via an SQL
WHERE.

Add unit tests as well.