content-sqlite: allow sqlite config modifications
Problem: In the near future the journal_mode and synchronous
configuration settings for sqlite may change while the module is
loaded.  The variables that store those settings do not allow
them to be changed.

Instead of storing sqlite configs as "const char *" types, change
types to "char *".  Use strdup() to store the configuration
setting, so that it can be freed/changed at a later time.