flux-exec: fix credits list update race
Problem: When a subprocess exits / fails, it is removed from the
"subprocess credits" list.  However, it is possible this exit / failure
is not seen yet before the subprocess credits is updated during a stdin
callback.  This can lead to the subprocess credits list calculating an
an invalid "min credits".  This can lead to the stdin callback being
started when it should not be.

Solution: If the number of "active" subprocesses does not equal the
number of subprocesses on the credits list, remove exited/failed credits
from the credits list in stdin_cb().

Fixes #6542