flux-job: support attach --tail
Problem: If flux job attach is executed on a long running job, it outputs
all job output for that job.  If the job output is extremely long, this
extremely long output can be very annoying to the user.  In many cases,
a user only wants to see the newest output (i.e. current job status, error
message causing the job to fail, etc.).

Support a --tail option.  By default it outputs the last 10 lines of output,
but an optional argument can be specified for additional lines.

Fixes #5035